深度ステンシル用の複数の定義を一つのファイルから個別のファイルに分離
◆リファクタリング内容
- Rasterizer_Setting_Inform.hに書かれていた定義をすべて個別のヘッダファイルへ移動
- 分離した定義をRasterizer_Setting_Inform.hからすべて読み込めるように変更

◆分離した定義
- E_Draw_Mode                  ラスタライザの表示モードの定義
- E_Mesh_Culling                面のカリングの基準の定義
- E_Mesh_Front                   時計回りと反時計回りのどちらを表面にするかの定義
- E_Antialiasing                   アンチエイリアシングの方法を記載
- S_Depth_Setting_Value    奥行きの値の設定を行うための定義
